pause input on focus loss

+++ EventListenerInterface.h
@@ -14,7 +14,7 @@ public:
     virtual void onWindowResize(int width, int height) {}
 
     // Callback for window resize event
-    virtual void onMouseMovement(double xpos, double ypos) {}
+    virtual void onMouseMovement(int xpos, int ypos) {}
 
 protected:
     bool firstMouse;
